Unify the two card renderers into a single shared implementation
src/render.ts (app) and src/render/cardSvg.ts (server/tests) were two
separate implementations of the same renderer that had to be kept in
sync by hand. The shared logic now lives in src/render/core.ts and both
files become thin entry points:

- src/render.ts re-exports the core renderer and keeps the browser-only
  helpers (buildFontCss, embedFontsInSvg, embedImagesInSvg)
- src/render/cardSvg.ts maps the historical server option names
  (fonts = binaries to inline as @font-face, fontSlots = slot → family)
  onto the core options (embedFonts / fonts) and re-exports the rest

The core merges the previously drifted behaviors:
- font handling: slot-map naming (app) and inline @font-face embedding
  (server) are both supported via separate option keys
- svgTextOnly and debug overlays (app-only) are now available to both
- image clip paths are hoisted into <defs> (server behavior)
- align/verticalAlign values are clamped to known values before being
  interpolated into style attributes (server behavior)
- renderLayoutSvg wireframes use the app styling (1.5 stroke,
  translucent fill); the server import of renderLayoutSvg was unused
- computeLayout keeps its historical per-entry semantics: mm-based in
  src/render.ts, px-based in src/render/cardSvg.ts

test/rendererParity.test.js now also asserts both entry points produce
byte-identical SVG for a layout exercising every item type, and
CLAUDE.md is updated to describe the unified architecture.
